[posterize] add edge detection endpoint + improve local dev (#897)
- Add Sobel edge detection algorithm to imagine library
  - Expose new /v1/imagine/edges endpoint in posterize service
  - Refactor posterize into modular structure (images, service, types)
  - Update Caddy configs to route edge detection and WebSocket endpoints
  - Add local development tooling (Caddyfile.local, local.sh, LOCAL.md)
  - Improve deployment documentation with service links
  - Update games_ws_backend to use versioned WebSocket paths
  - Add fallback handler and increase Bazel idle timeout
  - Fix gaussian blur kernel types (u8 -> i32) for signed convolution

+++ rust/imagine/src/radius.rs
@@ -5,7 +5,7 @@ pub enum Radius {
 }
 
 impl Radius {
-    pub fn gaussian_kernel(&self) -> &'static [u8] {
+    pub fn gaussian_kernel(&self) -> &'static [i32] {
         match self {
             Radius::Three => &[1, 3, 1, 3, 9, 3, 1, 3, 1],
             Radius::Five => &[
